verifyToken post good

+++ src/routes/post.js
@@ -4,6 +4,7 @@ const post = require("../models/post");
 const router = express.Router();
 const posts = require("../middleware/postFunc")
 const path = require('path');
+const token = require("../middleware/authTokens")
 const multer  = require('multer');
 
 const storage = require('../middleware/storage')
@@ -13,9 +14,15 @@ const app = express();
 //_________________Create______________
 router.post("/create", async (request, response) => {
     try {
-      const newPost = await posts.create(request)
-      await newPost.save()
-      response.json({status:"200",message:"Post creado correctamente"})
+      const verified = token.verify(request.headers['authorization'].split(" ")[1])
+      if(verified){
+        const newPost = await posts.create(request)
+        await newPost.save()
+        response.json({status:"200",message:"Post creado correctamente"})
+      }else{
+        response.json({status:403,message:"Forbidden"})
+      }
+      
     } catch (error) {
       console.log(error)
       response.json({status:"400" ,message:error});
@@ -25,8 +32,14 @@ router.post("/create", async (request, response) => {
 //_________________gets__________________
 router.get("/", async (request, response) => {
     try {
-      const posts = await post.find();
-      response.json({status:"200",data:posts});
+      const verified = token.verify(request.headers['authorization'].split(" ")[1])
+
+      if(verified){
+        const posts = await post.find();
+        response.json({status:"200",data:posts});
+      }else{
+        response.json({status:403,message:"Forbidden"})
+      }
     } catch (error) {
       response.json(error);
     }
@@ -35,8 +48,14 @@ router.get("/", async (request, response) => {
 router.get("/:id", async (request, response) => {
     const { id } = request.params;
     try {
+      const verified = token.verify(request.headers['authorization'].split(" ")[1])
+
+      if(verified){
         const postFinded = await post.findById(id);
         response.json({status:"ok",data:postFinded});
+      }else{
+        response.json({status:"403",message:"forbidden"})
+      }
     } catch (error) {
         response.json(error);
     }
@@ -44,9 +63,15 @@ router.get("/:id", async (request, response) => {
 
 router.post("/byuser", async (request, response) => {
   try {
+    const verified = token.verify(request.headers['authorization'].split(" ")[1])
+
+      if(verified){
     const{user} = request.body;
     const posts = await post.find({user:user});
     response.json({status:"200",data:posts});
+  }else{
+    response.json({status:"403",message:"forbidden"})
+  }
   } catch (error) {
     response.json(error);
   }
@@ -54,21 +79,39 @@ router.post("/byuser", async (request, response) => {
 
 //_________________________Update______________________
 router.put("/:id", async (request, response) => {
+
+  const verified = token.verify(request.headers['authorization'].split(" ")[1])
+  if(verified){
     await posts.update(request,response)
+  }else{
+    response.json({status:"403",message:"forbidden"})
+  }
 });
 
 router.put("/admupdate/:id", async (request, response) => {
+  const verified = token.verify(request.headers['authorization'].split(" ")[1])
+
+      if(verified){
   await posts.admUpdate(request,response)
+}else{
+  response.json({status:"403",message:"forbidden"})
+  
+}
 });
 
 //_____________________delete________________
 router.delete("/:id", async (request, response) => {
     const { id } = request.params;
     try {
+      const verified = token.verify(request.headers['authorization'].split(" ")[1])
+      if(verified){
         const removed = await post.deleteOne({ _id: id });
-        response.json(removed);
+        response.json({status:"200",data:removed});
+      }else{
+        response.json({status:"403",message:"forbidden"});
+      }
     } catch (error) {
-        response.json(error);
+        response.json({error:error});
     }
 });
 
@@ -77,7 +120,12 @@ router.delete("/:id", async (request, response) => {
 router.use('/img',express.static(path.join(__dirname,'../../img/post')));
 
 app.get('/img/:image', (req, res) => {
+  const verified = token.verify(request.headers['authorization'].split(" ")[1])
+  if(verified){
     res.sendFile(path.join(__dirname, 'img', req.params.image));
+  }else{
+    response.json({status:"403",message:"forbidden"})
+  }
   });
 
 
@@ -101,6 +149,10 @@ router.post('/img', upload.single('image'), async (req, res) => {
 
 router.put('/deleteimg/:id', async (req, res) => {
   try {
+    const header = req.headers['authorization']
+    console.log(header);
+    const verified = token.verify(header.split(" ")[1])
+  if(verified){
     const { id } = req.params;
     image = {
       image:"example-post.jpg"
@@ -113,6 +165,9 @@ router.put('/deleteimg/:id', async (req, res) => {
       }
     );
     res.json({status:"200",post:updated});
+  }else{
+    res.json({status:"403",message:"forbidden"})
+  }
   } catch (error) {
     res.json({status:"500",message:error});
     console.log(error)
